remove parquet_go18*
